refactor(core): one assertion vocabulary for the four conformance kits
`assert` was written out four times (index.ts, store-ops.ts, app-access.ts,
knowledge.ts), `assertDeepEqual` and `assertBytesEqual` twice each, and
`assertParses` twice — once zod-typed, once hand-typed to dodge the import.
store-ops.ts's own comment named the cause: "mirrors the ones in index.ts —
keep conformance self-contained", which is true of the barrel and not of a
sibling file.

They move to conformance/assertions.ts, deliberately NOT re-exported from
./index.js: `@vendoai/core/conformance`'s inventory is the kit surface an
adapter author consumes, not how the kits are written, and re-exporting
would make every kit a value-cycle with the barrel that re-exports it.
contract-coverage.e2e.test.ts's exact-inventory assertion is unchanged and
green.

The two `assertDeepEqual` copies had already diverged: store-ops' appended
the two canonical forms to the failure message and index.ts's did not. The
informative one wins, so a conformance failure in either kit now says what
differed. No case body changed.

core: 1071 pass. Downstream canary: packages/store's conformance.test.ts,
ops.conformance.test.ts and app-access.test.ts — 115 tests against real
PGlite, green.
